chore(runtime): move generic testing utils out of the `next/testing` dir
