+++ packages/runtime/src/next/testing/breakpoints.ts
@@ -1,3 +1,1 @@
-export const DESKTOP_MEDIA_QUERY = 'only screen and (min-width: 769px)'
-export const TABLET_MEDIA_QUERY = 'only screen and (min-width: 769px)'
-export const MOBILE_MEDIA_QUERY = 'only screen and (min-width: 769px)'
+export * from '../../testing/breakpoints'
